remove subsriber container
