+++ prince-archiver/compose.dev.yml
@@ -37,6 +37,7 @@ services:
   cron-worker:
     environment:
       <<: *aws-env-variables
+      RABBITMQ_DSN: amqp://guest:guest@rabbitmq:5672
       SURF_USERNAME: ""
       SURF_PASSWORD: ""
     volumes:
@@ -45,14 +46,6 @@ services:
     depends_on:
       s3:
         condition: service_healthy
-
-  data-archive-subscriber:
-    environment:
-      RABBITMQ_DSN: amqp://guest:guest@rabbitmq:5672
-    volumes:
-      - ./prince_archiver:/app/prince_archiver
-      - ./alembic/:/app/alembic
-    depends_on:
       rabbitmq:
         condition: service_healthy
