+++ prince-archiver/compose.prod.yml
@@ -39,21 +39,12 @@ services:
   cron-worker:
     environment:
       <<: *aws-env-variables
-      SURF_USERNAME: ${SURF_USERNAME}
-      SURF_PASSWORD: ${SURF_PASSWORD}
-      SENTRY_DSN: ${SENTRY_DSN}
-      UPLOAD_EXPIRY_DAYS: ${UPLOAD_EXPIRY_DAYS:-5}
       ARCHIVE_TRANSITION_DAYS: ${ARCHIVE_TRANSITION_DAYS:-2}
-      WEBHOOK_URL: ${WEBHOOK_URL}
-    labels: *log-labels
-    deploy:
-      restart_policy:
-        condition: on-failure
-
-  data-archive-subscriber:
-    environment:
       SENTRY_DSN: ${SENTRY_DSN}
       RABBITMQ_DSN: ${RABBITMQ_DSN}
+      SURF_USERNAME: ${SURF_USERNAME}
+      SURF_PASSWORD: ${SURF_PASSWORD}
+      WEBHOOK_URL: ${WEBHOOK_URL}
     labels: *log-labels
     deploy:
       restart_policy:
