+++ prince-archiver/prince_archiver/data_archive/subscriber.py
@@ -1,4 +1,3 @@
-import asyncio
 import logging
 from contextlib import AsyncExitStack
 from dataclasses import dataclass
@@ -6,14 +5,6 @@ from typing import Awaitable, Callable
 
 from aio_pika import ExchangeType, connect_robust
 from aio_pika.abc import AbstractIncomingMessage
-
-from prince_archiver.config import SubscriberSettings
-from prince_archiver.db import UnitOfWork, get_session_maker
-from prince_archiver.logging import configure_logging
-from prince_archiver.messagebus import MessageBus
-
-from .dto import UpdateArchiveEntries
-from .handlers import SubscriberMessageHandler, add_data_archive_entries
 
 LOGGER = logging.getLogger(__name__)
 
@@ -55,34 +46,10 @@ class ManagedSubscriber:
         await queue.bind(exchange)
         await queue.consume(self.message_handler)
 
+        LOGGER.info("Subscribed to channel")
+
         return self
 
-    async def __aexit__(self, *_):
+    async def __aexit__(self, exc_type, exc_value, exc_tb):
         if self.exit_stack:
             await self.exit_stack.aclose()
-
-
-async def main(*, _settings: SubscriberSettings | None = None):
-    configure_logging()
-
-    settings = _settings or SubscriberSettings()
-
-    sessionmaker = get_session_maker(str(settings.POSTGRES_DSN))
-
-    def _messagebus_factory() -> MessageBus:
-        return MessageBus(
-            handlers={UpdateArchiveEntries: [add_data_archive_entries]},
-            uow=UnitOfWork(sessionmaker),
-        )
-
-    subscriber = ManagedSubscriber(
-        connection_url=settings.RABBITMQ_DSN,
-        message_handler=SubscriberMessageHandler(_messagebus_factory),
-    )
-    async with subscriber:
-        logging.info("Starting up")
-        await asyncio.Future()
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
