+++ prince-archiver/prince_archiver/data_archive/worker.py
@@ -14,9 +14,16 @@ from prince_archiver.config import ArchiveWorkerSettings
 from prince_archiver.db import UnitOfWork, get_session_maker
 from prince_archiver.file import managed_file_system
 from prince_archiver.logging import configure_logging
+from prince_archiver.messagebus import MessageBus
 
 from .archiver import AbstractArchiver, Settings, SurfArchiver
+from .dto import UpdateArchiveEntries
+from .handlers import (
+    SubscriberMessageHandler,
+    add_data_archive_entries,
+)
 from .reporter import Messenger, Reporter
+from .subscriber import ManagedSubscriber
 
 LOGGER = logging.getLogger(__name__)
 
@@ -55,10 +62,27 @@ async def startup(ctx: dict):
     exit_stack = await AsyncExitStack().__aenter__()
     settings = ArchiveWorkerSettings()
 
-    sessionmaker = get_session_maker(str(settings.POSTGRES_DSN))
+    uow_factory = partial(
+        UnitOfWork,
+        get_session_maker(str(settings.POSTGRES_DSN)),
+    )
+
+    def _messagebus_factory() -> MessageBus:
+        return MessageBus(
+            handlers={UpdateArchiveEntries: [add_data_archive_entries]},
+            uow=uow_factory(),
+        )
 
     ctx["exit_stack"] = exit_stack
     ctx["settings"] = settings
+    ctx["uow_factory"] = uow_factory
+
+    # Configure archive subscriber
+    subscriber = ManagedSubscriber(
+        connection_url=settings.RABBITMQ_DSN,
+        message_handler=SubscriberMessageHandler(_messagebus_factory),
+    )
+    await exit_stack.enter_async_context(subscriber)
 
     # Configure archiver
     if settings.SURF_USERNAME and settings.SURF_PASSWORD:
@@ -78,7 +102,6 @@ async def startup(ctx: dict):
         messenger = Messenger(client, str(settings.WEBHOOK_URL))
     s3 = await exit_stack.enter_async_context(managed_file_system(settings))
 
-    ctx["uow_factory"] = partial(UnitOfWork, sessionmaker)
     ctx["reporter"] = Reporter(s3, messenger=messenger)
 
     LOGGER.info("Start up complete")
